PLCWebConnect/RaspberryPi/requirements.txt
flask>=3.0.0
pymodbus>=3.6.0
pyserial>=3.5
RPi.GPIO>=0.7.1
gpiozero>=1.6.2